feat: add write report

+++ include/aho_queue.h
@@ -2,6 +2,7 @@
 #define AHO_QUEUE_H
 
 #include <stdbool.h>
+#include <stdint.h>
 #include "aho_config.h"
 
 /**
@@ -16,10 +17,10 @@
  * Utilizada para percorrer o trie/autômato na construção dos links de falha.
  */
 typedef struct {
-    int queue[AC_MAX_VERTICES]; /**< Vetor circular de índices dos nós. */
-    int head;                   /**< Índice do início da fila. */
-    int tail;                   /**< Índice do final da fila. */
-    int count;                  /**< Número de elementos atualmente na fila. */
+    uint8_t queue[AC_MAX_VERTICES]; /**< Vetor circular de índices dos nós (1 byte cada). */
+    uint8_t head;                   /**< Índice do início da fila. */
+    uint8_t tail;                   /**< Índice do final da fila. */
+    uint8_t count;                  /**< Número de elementos atualmente na fila. */
 } aho_queue_t;
 
 /**
@@ -34,14 +35,14 @@ void aho_queue_init(aho_queue_t *q);
  * @param node_idx Índice do nó a ser enfileirado.
  * @return true se a operação foi bem-sucedida, false se a fila estiver cheia.
  */
-bool aho_queue_enqueue(aho_queue_t *q, int node_idx);
+bool aho_queue_enqueue(aho_queue_t *q, uint8_t node_idx);
 
 /**
  * @brief Remove e retorna o índice do nó no início da fila.
  * @param q Ponteiro para a fila.
- * @return Índice do nó desenfileirado, ou -1 se a fila estiver vazia.
+ * @return Índice do nó desenfileirado, ou 255 se a fila estiver vazia.
  */
-int aho_queue_dequeue(aho_queue_t *q);
+uint8_t aho_queue_dequeue(aho_queue_t *q);
 
 /**
  * @brief Verifica se a fila está vazia.
